Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/compile-error-same-origin-with-hash-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected token ')'

PASS window.onerror - compile error in <script src=...> with hash
PASS window.onerror - compile error in <script src=...> with hash (column)
